requirements.txt
fastapi
uvicorn
python-multipart
deepface
weaviate-client
numpy
tf-keras
aiofiles